reoganize schema into a directory
